requirements.txt
FastMCP
httpx
pydantic
openai
loadenv
geopy
